+++ test/tools/jsontests/BlockChainTests.h
@@ -21,8 +21,9 @@ along with cpp-ethereum.  If not, see <http://www.gnu.org/licenses/>.
 #pragma once
 #include <libethashseal/Ethash.h>
 #include <libethashseal/GenesisInfo.h>
-#include <test/tools/libtesteth/TestSuite.h>
+#include <test/tools/jsontests/StateTestFixtureBase.h>
 #include <test/tools/libtesteth/BlockChainHelper.h>
+#include <test/tools/libtesteth/TestSuite.h>
 #include <boost/filesystem/path.hpp>
 
 using namespace dev;
@@ -51,6 +52,12 @@ class BCGeneralStateTestsSuite : public BlockchainValidTestSuite
 {
     boost::filesystem::path suiteFolder() const override;
     boost::filesystem::path suiteFillerFolder() const override;
+};
+
+class bcGeneralTestsFixture : public StateTestFixtureBase<BCGeneralStateTestsSuite>
+{
+public:
+    bcGeneralTestsFixture() : StateTestFixtureBase({TestExecution::RequireOptionAll}) {}
 };
 
 class TransitionTestsSuite: public TestSuite